requirements.txt
pyyaml
pytest
docker-compose
